refactor: replacing Koin by Dagger
