+++ app/src/main/java/me/loterio/randomemoji/EmojisApplication.kt
@@ -1,24 +1,27 @@
 package me.loterio.randomemoji
 
 import android.app.Application
-import me.loterio.randomemoji.di.networkModule
-import me.loterio.randomemoji.di.repositoryModule
-import me.loterio.randomemoji.presentation.emojiListViewModel
-import me.loterio.randomemoji.presentation.fragmentModule
-import org.koin.android.ext.koin.androidContext
-import org.koin.android.ext.koin.androidLogger
-import org.koin.core.context.startKoin
+import me.loterio.randomemoji.di.AndroidCoreModule
+import me.loterio.randomemoji.di.AppComponent
+import me.loterio.randomemoji.di.DaggerAppComponent
 
 class EmojisApplication : Application() {
 
+
+    private lateinit var appComponent: AppComponent
+
     override fun onCreate() {
         super.onCreate()
-        startKoin {
-            androidLogger()
-            androidContext(this@EmojisApplication)
-            modules(
-                networkModule, repositoryModule, emojiListViewModel, fragmentModule
-            )
-        }
+
+        appComponent = DaggerAppComponent.builder()
+            .androidCoreModule(AndroidCoreModule())
+            .build()
+
+        application = this
     }
+
+    companion object  {
+        lateinit var application: EmojisApplication
+    }
+
 }
