+++ app/src/main/java/me/loterio/randomemoji/di/AndroidCoreModule.kt
@@ -1,0 +1,15 @@
+package me.loterio.randomemoji.di
+
+import android.content.Context
+import dagger.Module
+import dagger.Provides
+import me.loterio.randomemoji.EmojisApplication
+import javax.inject.Singleton
+
+@Module
+class AndroidCoreModule() {
+
+    @Singleton
+    @Provides
+    fun provideApplicationContext(): Context = EmojisApplication.application
+}
