+++ app/src/main/java/me/loterio/randomemoji/di/AppComponent.kt
@@ -1,0 +1,15 @@
+package me.loterio.randomemoji.di
+
+import dagger.Component
+import me.loterio.randomemoji.presentation.FragmenEmojiList
+import javax.inject.Singleton
+
+@Singleton
+@Component(modules = [
+    AndroidCoreModule::class,
+    NetworkModule::class,
+    RepositoryModule::class])
+interface AppComponent {
+
+    fun inject(fragment: FragmenEmojiList)
+}
