+++ app/src/main/java/me/loterio/randomemoji/presentation/FragmenEmojiList.kt
@@ -1,26 +1,19 @@
 package me.loterio.randomemoji.presentation
 
-import android.animation.Animator
-import android.animation.AnimatorListenerAdapter
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.databinding.DataBindingUtil
-import androidx.databinding.Observable
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.Observer
 import androidx.recyclerview.widget.GridLayoutManager
 import me.loterio.randomemoji.R
-import me.loterio.randomemoji.core.log
 import me.loterio.randomemoji.databinding.FragmentEmojiListBinding
-import me.loterio.randomemoji.model.Emoji
+import me.loterio.randomemoji.di.DaggerAppComponent
+import me.loterio.randomemoji.domain.model.Emoji
 import org.koin.android.viewmodel.ext.android.viewModel
-import org.koin.dsl.module
 
-val fragmentModule = module {
-    factory { FragmenEmojiList() }
-}
 class FragmenEmojiList: Fragment() {
     private var shortAnimationDuration: Int = 0
     private val emojiListViewModel: EmojiListViewModel by viewModel()
@@ -31,6 +24,10 @@ class FragmenEmojiList: Fragment() {
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
+
+        //DaggerAppComponent.factory().create(applicationContext = context)
+
+
         binding  = DataBindingUtil.inflate(inflater,
             R.layout.fragment_emoji_list, container, false)
         binding.lifecycleOwner = this
